move supabaseAdmin function to server side
