+++ src/lib/supabase.ts
@@ -1,18 +1,5 @@
-import { createClient } from "@supabase/supabase-js";
 import { createBrowserClient } from "@supabase/ssr";
 import { getSupabaseJwt } from "@/utils/actions";
-
-const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL || '';
-// For admin usage
-const supabaseServiceRoleKey = process.env.SUPABASE_SERVICE_ROLE_KEY || '';
-
-// Create a Supabase client with the service role key for server-side operations
-export const supabaseAdmin = createClient(supabaseUrl, supabaseServiceRoleKey, {
-  auth: {
-    autoRefreshToken: false,
-    persistSession: false,
-  },
-});
 
 // Create a Supabase client for client-side operations
 export const createSupabaseClient = () => {
